Move `PredefinedGitHubResponses` to `testutil-github`.
